Remove the telemetry data parameter from the public API
There was a parameter to set the telemetry data in the
constructor of the client. A normal user should not be
able to specify the format or the content of the telemetry
data we sent to server. Therefore, I have removed that
argument from the client's constructor.

Also, the implementation was constructing the headers for
each HTTP call, which is unnecessary. Now, we create the
headers once in the client's constructor and reuse it.

Also, removed the unnecessary modularazitation of the
http folder, and replaced that with a single file, and improved
its tests.
