+++ upstash_redis/client.py
@@ -1,21 +1,19 @@
 from os import environ
 from typing import Any, List, Type, Union
 
-from aiohttp import ClientSession
+from requests import Session
+
 from upstash_redis.commands.commands import Commands
 from upstash_redis.config import (
+    ALLOW_TELEMETRY,
+    FORMAT_RETURN,
     REST_ENCODING,
     REST_RETRIES,
     REST_RETRY_INTERVAL,
-    FORMAT_RETURN,
-    ALLOW_TELEMETRY,
 )
-from upstash_redis.http.execute import sync_execute
-
+from upstash_redis.http import make_headers, sync_execute
 from upstash_redis.schema.http import RESTEncoding, RESTResult
-from upstash_redis.schema.telemetry import TelemetryData
 from upstash_redis.utils.format import FormattedResponse
-from requests import Session
 
 
 class Redis(FormattedResponse, Commands):
@@ -28,7 +26,6 @@ class Redis(FormattedResponse, Commands):
         rest_retry_interval: int = REST_RETRY_INTERVAL,  # Seconds.
         format_return: bool = FORMAT_RETURN,
         allow_telemetry: bool = ALLOW_TELEMETRY,
-        telemetry_data: Union[TelemetryData, None] = None,
     ):
         """
         :param url: UPSTASH_REDIS_REST_URL in the console
@@ -51,9 +48,9 @@ class Redis(FormattedResponse, Commands):
         self.rest_retries = rest_retries
         self.rest_retry_interval = rest_retry_interval
 
-        self.telemetry_data = telemetry_data
         self.FORMATTERS = self.__class__.FORMATTERS
 
+        self._headers = make_headers(token, rest_encoding, allow_telemetry)
         self._session = Session()
 
     def __enter__(self):
@@ -86,7 +83,6 @@ class Redis(FormattedResponse, Commands):
         rest_retry_interval: int = REST_RETRY_INTERVAL,
         format_return: bool = FORMAT_RETURN,
         allow_telemetry: bool = ALLOW_TELEMETRY,
-        telemetry_data: Union[TelemetryData, None] = None,
     ):
         """
         Load the credentials from environment.
@@ -106,7 +102,6 @@ class Redis(FormattedResponse, Commands):
             rest_retry_interval,
             format_return,
             allow_telemetry,
-            telemetry_data,
         )
 
     def close(self):
@@ -124,13 +119,11 @@ class Redis(FormattedResponse, Commands):
         res = sync_execute(
             session=self._session,
             url=self.url,
-            token=self.token,
+            headers=self._headers,
             encoding=self.rest_encoding,
             retries=self.rest_retries,
             retry_interval=self.rest_retry_interval,
             command=command,
-            allow_telemetry=self.allow_telemetry,
-            telemetry_data=self.telemetry_data,
         )
 
         main_command = command[0]
