--- upstash_redis/schema/telemetry.py
@@ -1,7 +1,0 @@
-from typing import TypedDict
-
-
-class TelemetryData(TypedDict, total=False):  # Allow omitting properties.
-    runtime: str
-    sdk: str
-    platform: str
